+++ include/utils/Benchmark.hpp
@@ -97,7 +97,7 @@ public:
 class InstrumentationTimer
 {
 public:
-    InstrumentationTimer(const char *name)
+    explicit InstrumentationTimer(const char *name)
         : m_Name(name), m_Stopped(false)
     {
         m_StartTimepoint = std::chrono::high_resolution_clock::now();