+++ include/utils/Expected.hpp
@@ -28,8 +28,8 @@ protected:
     Expected() {}
 
 public:
-    Expected(T const &r) : result{r}, gotResult{true} {}
-    Expected(T &&r) : result{std::move(r)}, gotResult{true} {}
+    explicit Expected(T const &r) : result{r}, gotResult{true} {}
+    explicit Expected(T &&r) : result{std::move(r)}, gotResult{true} {}
 
     Expected(Expected const &e) : gotResult{e.gotResult}
     {
@@ -48,7 +48,7 @@ public:
     }
 
     template <typename E>
-    Expected<T>(E const &e) : spam{std::make_exception_ptr(e)}, gotResult{false} {}
+    explicit Expected<T>(E const &e) : spam{std::make_exception_ptr(e)}, gotResult{false} {}
 
     ~Expected() {}
 
@@ -173,10 +173,10 @@ class Expected<void>
 
 public:
     template <typename E>
-    Expected(E const &e) : spam{std::make_exception_ptr(e)} {}
+    explicit Expected(E const &e) : spam{std::make_exception_ptr(e)} {}
 
     template <typename T>
-    Expected(const Expected<T> &e)
+    explicit Expected(const Expected<T> &e)
     {
         if (!e.gotResult)
             new (&spam) std::exception_ptr{e.spam};